Add initial pattern for tests in cmd package
I've successfully used a pattern similar to this on previous projects.
This approach allows testing the command in-process and stubbing out
stdin, stdout, and stderr.

There are some filesystem interactions, so these tests ought to be run
with t.Parallel, but right now there is a data race around some global
state in commands. We can fix the race later and then make the tests run
concurrently.